requirements.txt
python-telegram-bot==21.4
yt-dlp>=2024.07.07
python-dotenv>=1.0.1
tqdm>=4.66.4
